+++ lib/auth/keystore/keystore_test.go
@@ -429,12 +429,19 @@ func newTestPack(ctx context.Context, t *testing.T) *testPack {
 	})
 
 	if config, ok := softHSMTestConfig(t); ok {
-		config.PKCS11.HostUUID = hostUUID
-		backend, err := newPKCS11KeyStore(&config.PKCS11, logger)
+		cfg := PKCS11Config{
+			Path:       config.PKCS11.Path,
+			SlotNumber: config.PKCS11.SlotNumber,
+			TokenLabel: config.PKCS11.TokenLabel,
+			Pin:        config.PKCS11.Pin,
+			HostUUID:   hostUUID,
+		}
+
+		backend, err := newPKCS11KeyStore(&cfg, logger)
 		require.NoError(t, err)
 		backends = append(backends, &backendDesc{
 			name:            "softhsm",
-			config:          config,
+			config:          Config{PKCS11: cfg},
 			backend:         backend,
 			expectedKeyType: types.PrivateKeyType_PKCS11,
 			unusedRawKey:    unusedPKCS11Key,
@@ -442,12 +449,19 @@ func newTestPack(ctx context.Context, t *testing.T) *testPack {
 	}
 
 	if config, ok := yubiHSMTestConfig(t); ok {
-		config.PKCS11.HostUUID = hostUUID
-		backend, err := newPKCS11KeyStore(&config.PKCS11, logger)
+		cfg := PKCS11Config{
+			Path:       config.PKCS11.Path,
+			SlotNumber: config.PKCS11.SlotNumber,
+			TokenLabel: config.PKCS11.TokenLabel,
+			Pin:        config.PKCS11.Pin,
+			HostUUID:   hostUUID,
+		}
+
+		backend, err := newPKCS11KeyStore(&cfg, logger)
 		require.NoError(t, err)
 		backends = append(backends, &backendDesc{
 			name:            "yubihsm",
-			config:          config,
+			config:          Config{PKCS11: cfg},
 			backend:         backend,
 			expectedKeyType: types.PrivateKeyType_PKCS11,
 			unusedRawKey:    unusedPKCS11Key,
@@ -455,12 +469,18 @@ func newTestPack(ctx context.Context, t *testing.T) *testPack {
 	}
 
 	if config, ok := cloudHSMTestConfig(t); ok {
-		config.PKCS11.HostUUID = hostUUID
-		backend, err := newPKCS11KeyStore(&config.PKCS11, logger)
+		cfg := PKCS11Config{
+			Path:       config.PKCS11.Path,
+			SlotNumber: config.PKCS11.SlotNumber,
+			TokenLabel: config.PKCS11.TokenLabel,
+			Pin:        config.PKCS11.Pin,
+			HostUUID:   hostUUID,
+		}
+		backend, err := newPKCS11KeyStore(&cfg, logger)
 		require.NoError(t, err)
 		backends = append(backends, &backendDesc{
 			name:            "yubihsm",
-			config:          config,
+			config:          Config{PKCS11: cfg},
 			backend:         backend,
 			expectedKeyType: types.PrivateKeyType_PKCS11,
 			unusedRawKey:    unusedPKCS11Key,
@@ -468,12 +488,17 @@ func newTestPack(ctx context.Context, t *testing.T) *testPack {
 	}
 
 	if config, ok := gcpKMSTestConfig(t); ok {
-		config.GCPKMS.HostUUID = hostUUID
-		backend, err := newGCPKMSKeyStore(ctx, &config.GCPKMS, logger)
+		cfg := GCPKMSConfig{
+			KeyRing:         config.GCPKMS.KeyRing,
+			ProtectionLevel: config.GCPKMS.ProtectionLevel,
+			HostUUID:        hostUUID,
+		}
+
+		backend, err := newGCPKMSKeyStore(ctx, &cfg, logger)
 		require.NoError(t, err)
 		backends = append(backends, &backendDesc{
 			name:            "gcp_kms",
-			config:          config,
+			config:          Config{GCPKMS: cfg},
 			backend:         backend,
 			expectedKeyType: types.PrivateKeyType_GCP_KMS,
 			unusedRawKey: gcpKMSKeyID{
@@ -504,12 +529,18 @@ func newTestPack(ctx context.Context, t *testing.T) *testPack {
 	})
 
 	if config, ok := awsKMSTestConfig(t); ok {
-		config.AWSKMS.clock = clock
-		backend, err := newAWSKMSKeystore(ctx, &config.AWSKMS, logger)
+		cfg := AWSKMSConfig{
+			Cluster:    config.AWSKMS.Cluster,
+			AWSAccount: config.AWSKMS.AWSAccount,
+			AWSRegion:  config.AWSKMS.AWSRegion,
+			clock:      clock,
+		}
+
+		backend, err := newAWSKMSKeystore(ctx, &cfg, logger)
 		require.NoError(t, err)
 		backends = append(backends, &backendDesc{
 			name:            "aws_kms",
-			config:          config,
+			config:          Config{AWSKMS: cfg},
 			backend:         backend,
 			expectedKeyType: types.PrivateKeyType_AWS_KMS,
 			unusedRawKey: awsKMSKeyID{
