+++ lib/srv/discovery/common/renaming_test.go
@@ -506,7 +506,7 @@ func makeEKSKubeCluster(t *testing.T, name, region, accountID, overrideLabel str
 			overrideLabel: aws.String(name),
 		},
 	}
-	kubeCluster, err := services.NewKubeClusterFromAWSEKS(aws.StringValue(eksCluster.Name), aws.StringValue(eksCluster.Arn), eksCluster.Tags)
+	kubeCluster, err := NewKubeClusterFromAWSEKS(aws.StringValue(eksCluster.Name), aws.StringValue(eksCluster.Arn), eksCluster.Tags)
 	require.NoError(t, err)
 	require.True(t, kubeCluster.IsAWS())
 	return kubeCluster
@@ -525,7 +525,7 @@ func makeAKSKubeCluster(t *testing.T, name, location, group, subID string) types
 		},
 		Properties: azure.AKSClusterProperties{},
 	}
-	kubeCluster, err := services.NewKubeClusterFromAzureAKS(aksCluster)
+	kubeCluster, err := NewKubeClusterFromAzureAKS(aksCluster)
 	require.NoError(t, err)
 	require.True(t, kubeCluster.IsAzure())
 	return kubeCluster
@@ -543,7 +543,7 @@ func makeGKEKubeCluster(t *testing.T, name, location, projectID string) types.Ku
 		Description: "desc1",
 	}
 
-	kubeCluster, err := services.NewKubeClusterFromGCPGKE(gkeCluster)
+	kubeCluster, err := NewKubeClusterFromGCPGKE(gkeCluster)
 	require.NoError(t, err)
 	require.True(t, kubeCluster.IsGCP())
 	return kubeCluster
